refactor: move BotStatus logic to a module
Closes #62
